requirements.txt
Flask
gunicorn
azure-cognitiveservices-speech
azure-ai-vision-imageanalysis
werkzeug
pydub
opencv-python
